+++ src/main/vuejs/components/about.vue
@@ -1,5 +1,8 @@
 <template>
   <div>
     <h1>About...</h1>
+    <div>
+      <button class="btn btn-danger">Click me now</button>
+    </div>
   </div>
 </template>