+++ src/main/vuejs/components/layout/page-footer.vue
@@ -1,0 +1,7 @@
+<template>
+  <footer class="footer">
+    <div class="text-xs-center container">
+      <span class="text-muted">Vue.js java starter Vopyright &copy; Tomasz Kucharzyk 2016</span>
+    </div>
+  </footer>
+</template>